+++ src/commands/moveToFinished-8.lua
@@ -44,10 +44,11 @@
 local rcall = redis.call
 
 -- Includes
---- @include "includes/updateParentDepsIfNeeded"
 --- @include "includes/destructureJobKey"
+--- @include "includes/moveJobFromWaitToActive"
 --- @include "includes/removeJob"
 --- @include "includes/trimEvents"
+--- @include "includes/updateParentDepsIfNeeded"
 
 local jobIdKey = KEYS[3]
 if rcall("EXISTS", jobIdKey) == 1 then -- // Make sure job exists
@@ -154,10 +155,7 @@ if rcall("EXISTS", jobIdKey) == 1 then -- // Make sure job exists
                 rcall("SET", lockKey, ARGV[10], "PX", ARGV[11])
             end
 
-            rcall("ZREM", KEYS[5], jobId) -- remove from priority
-            rcall("XADD", KEYS[6], "*", "event", "active", "jobId", jobId,
-                  "prev", "waiting")
-            rcall("HSET", jobKey, "processedOn", timestamp)
+            moveJobFromWaitToActive(KEYS[5], KEYS[6], jobKey, jobId, timestamp)
 
             return {rcall("HGETALL", jobKey), jobId} -- get job data
         else