+++ src/types/finished-status.ts
@@ -1,3 +1,3 @@
-export type FinishedTarget = 'completed' | 'failed';
+export type FinishedStatus = 'completed' | 'failed';
 
 export type FinishedPropValAttribute = 'returnvalue' | 'failedReason';